Refactor some parts
