+++ TechnoFit/TechnoFitQtClient/Sources/main_test.cpp
@@ -1,8 +1,6 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 
-
-
 int main(int argc, char** argv)
 {
     testing::InitGoogleTest(&argc, argv);
