+++ TechnoFit/TechnoFitServer/Database/DataBaseDriver.cpp
@@ -22,9 +22,7 @@ std::string DataBaseDriver::get(std::string query)
     std::string pulse = std::to_string(distribution_for_pulse(generator));
     std::string O2 = std::to_string(distribution_for_O2(generator));
     std::string temprature = std::to_string(distribution_for_temprature(generator));
-    // получает запрос как входной параметр делает запрос отдаёт репозиторию результат
     std::string reply = pulse + " " + O2 + " " + temprature;
-    // в клиенте паршу это как 3 чиселки подряд первая кислород, вторая пульс, третья температура
     return reply;
 }
 void DataBaseDriver::save()
