+++ TechnoFit/TechnoFitServer/main.cpp
@@ -1,6 +1,6 @@
 #include "global.h"
 #include "httplistener.h"
-#include "Handlers/RequestHandler.h"
+#include "Handlers/RequestRouter.h"
 #include "logger.h"
 #include <QCoreApplication>
 #include <QDir>
